Rename components folder to trackedComponents for consistency
