+++ tracker/trackers/react-native/src/trackedComponents/TrackedModal.tsx
