+++ tracker/trackers/react-native/src/trackedComponents/TrackedText.tsx
