+++ tracker/trackers/react-native/src/trackedComponents/TrackedTouchableHighlight.tsx
