+++ tracker/trackers/react-native/src/trackedComponents/TrackedTouchableOpacity.tsx
